# KR2/fbr-practice_9_10/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FBR Practice 9 Backend API",
    "version": "1.0.0",
    "description": "Backend API for practices 9 and 10"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-refresh-token"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "first_name", "last_name", "password"],
        "properties": {
          "email": {
            "type": "string",
            "example": "evgeniya@mail.com"
          },
          "first_name": {
            "type": "string",
            "example": "Evgeniya"
          },
          "last_name": {
            "type": "string",
            "example": "Razina"
          },
          "password": {
            "type": "string",
            "example": "123"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "example": "evgeniya@mail.com"
          },
          "password": {
            "type": "string",
            "example": "123"
          }
        }
      },
      "CreateProductRequest": {
        "type": "object",
        "required": ["title", "category", "description", "price"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Lego"
          },
          "category": {
            "type": "string",
            "example": "Lego Technic"
          },
          "description": {
            "type": "string",
            "example": "Конструктор для детей от 14 лет"
          },
          "price": {
            "type": "number",
            "example": 99999
          }
        }
      },
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Lego Star Wars"
          },
          "category": {
            "type": "string",
            "example": "Lego Technic"
          },
          "description": {
            "type": "string",
            "example": "Конструктор для детей от 10 лет"
          },
          "price": {
            "type": "number",
            "example": 109999
          }
        }
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Регистрация пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь создан"
          },
          "400": {
            "description": "Некорректные данные"
          },
          "409": {
            "description": "Пользователь уже существует"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Вход в систему",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход"
          },
          "401": {
            "description": "Неверные данные"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Обновление пары токенов",
        "security": [
          {
            "refreshHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Новая пара токенов"
          },
          "401": {
            "description": "Refresh token отсутствует"
          },
          "403": {
            "description": "Refresh token недействителен"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Получить текущего пользователя",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Данные пользователя"
          },
          "401": {
            "description": "Требуется access token"
          }
        }
      }
    },
    "/api/products": {
      "post": {
        "summary": "Создать товар",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Товар создан"
          },
          "400": {
            "description": "Некорректные данные"
          }
        }
      },
      "get": {
        "summary": "Получить список товаров",
        "responses": {
          "200": {
            "description": "Список товаров"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Получить товар по id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар найден"
          },
          "401": {
            "description": "Требуется access token"
          },
          "404": {
            "description": "Товар не найден"
          }
        }
      },
      "put": {
        "summary": "Обновить товар",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Товар обновлён"
          },
          "401": {
            "description": "Требуется access token"
          },
          "404": {
            "description": "Товар не найден"
          }
        }
      },
      "delete": {
        "summary": "Удалить товар",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар удалён"
          },
          "401": {
            "description": "Требуется access token"
          },
          "404": {
            "description": "Товар не найден"
          }
        }
      }
    }
  }
}